+++ src/senaite/core/content/storagelocations.py
@@ -18,25 +18,21 @@
 # Copyright 2018-2024 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
-from bika.lims.config import PROJECTNAME
-from bika.lims.interfaces import IStorageLocations
-from plone.app.folder.folder import ATFolder
-from plone.app.folder.folder import ATFolderSchema
-from Products.Archetypes import atapi
-from Products.ATContentTypes.content import schemata
+from bika.lims.interfaces import IDoNotSupportSnapshots
+from plone.dexterity.content import Container
+from plone.supermodel import model
 from senaite.core.interfaces import IHideActionsMenu
-from zope.interface.declarations import implements
+from senaite.core.interfaces import IStorageLocations
+from zope.interface import implementer
 
 
-schema = ATFolderSchema.copy()
+class IStorageLocationsSchema(model.Schema):
+    """Schema interface
+    """
 
 
-# TODO: Migrated to DX - https://github.com/senaite/senaite.core/pull/2550
-class StorageLocations(ATFolder):
-    implements(IStorageLocations, IHideActionsMenu)
-    displayContentsTab = False
-    schema = schema
-
-
-schemata.finalizeATCTSchema(schema, folderish=True, moveDiscussion=False)
-atapi.registerType(StorageLocations, PROJECTNAME)
+@implementer(IStorageLocations, IStorageLocationsSchema,
+             IDoNotSupportSnapshots, IHideActionsMenu)
+class StorageLocations(Container):
+    """A container for storage locations
+    """