+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.6.0 (unreleased)
 ------------------
 
+- #2550 Migrate StorageLocations to Dexterity
 - #2545 Migrate Sub Groups to Dexterity
 - #2556 Fix adding Instrument into Manufacturer view
 - #2549 Fix contact fullname is not updated in samples listing after edition